requirements.txt
opencv-contrib-python==4.4.0.46
numpy=1.19.0
matplotlib
requests
Pillow
python-levenshtein